requirements.txt
streamlit
langchain
faiss-cpu
pypdf
python-dotenv
google-generativeai
